added config example and reorganized
