+++ app.js
@@ -7,11 +7,11 @@ client.login(apikey);
 
 
 const {music} = require("./commands/music/music");
-
+const queue = [];
 
 client.on("ready", () => {
     console.log("Bot Launched... ")
 })
 
 
-music(client);
+music(client, queue);
