+++ config.example.js
@@ -1,0 +1,5 @@
+module.exports = {
+    apikey: "your discord api key here",
+    prefix: 'your choosen prefix here',
+    googleapikey: "your google api key here"
+}
